+++ NEWS.md
@@ -1,6 +1,13 @@
+# CGGP 1.0.2
+
+* Fixed errors in examples.
+
 # CGGP 1.0.1
 
-* Change default correlation function to power exponential.
+* Changed default correlation function to power exponential,
+which we have found to be more numerically stable.
+
+* Accepted to CRAN on 6/16/2019.
 
 # CGGP 1.0.0